Add ability to schedule when the kiln starts
This adds the possibility to use a datepicker in the modal after
clicking the start button to schedule when the kiln should start
running by itself automatically.
The timer is implemented in the backend and the start is triggered
there so closing or refreshing the browser does not stop it.
In the state when it's "waiting to start", the frontend state
changes so that the glowing timer icon is now shown instead of the
previously unused door icon. The state is also displayed as
SCHEDULED and above it the info states when it's due to start:
"Start at: ..."